Removed some hardcode comments and adjust file for pre-computation
